+++ akshare/futures_derivative/futures_index_return_nh.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python
 # -*- coding:utf-8 -*-
 """
-Date: 2023/5/31 16:19
+Date: 2023/11/17 18:20
 Desc: 南华期货-商品指数历史走势-收益率指数-数值
 https://www.nanhua.net/nhzc/varietytrend.html
 1000 点开始, 用收益率累计
@@ -35,7 +35,7 @@ def futures_return_index_nh(symbol: str = "Y") -> pd.DataFrame:
         data_json = r.json()
         temp_df = pd.DataFrame(data_json)
         temp_df.columns = ["date", "value"]
-        temp_df["date"] = pd.to_datetime(temp_df["date"], unit="ms")
+        temp_df["date"] = pd.to_datetime(temp_df["date"], unit="ms", errors="coerce")
         temp_df["date"] = temp_df["date"].dt.tz_localize("UTC")
         temp_df["date"] = temp_df["date"].dt.tz_convert("Asia/Shanghai").dt.date
         temp_df["value"] = pd.to_numeric(temp_df["value"], errors="coerce")