+++ docs/changelog.md
@@ -70,6 +70,13 @@
 
 ## 更新说明详情
 
+1.11.76 fix: fix option_czce_daily interface
+
+    1. 修复 option_czce_daily 接口
+    2. 移除 stock_hot_tgb 接口
+    3. 修复 futures_board_index_nh 接口
+    4. 修复 stock_hk_hist 接口
+
 1.11.75 fix: fix futures_comex_inventory interface
 
     1. 修复 futures_comex_inventory 接口
@@ -3144,6 +3151,8 @@
     3. 修改 air_quality_hist 接口的请求日期格式，从 '2022-01-01' 为 '20220101'
 
 ## 版本更新说明
+
+1.11.76 fix: fix option_czce_daily interface
 
 1.11.75 fix: fix futures_comex_inventory interface